Updated UI stuff

+++ src/Entities/UI/Button.h
@@ -7,8 +7,8 @@ public:
 
     void Update() override;
     void Draw() override;
-    inline void MoveBy(Vec2f offset) override { Entity::MoveBy(offset); Move(); }
-    inline void MoveTo(Vec2f new_pos) override { Entity::MoveTo(new_pos); Move(); }
+    inline void MoveBy(Vec2i offset) override { Entity::MoveBy(offset); Move(); }
+    inline void MoveTo(Vec2i new_pos) override { Entity::MoveTo(new_pos); Move(); }
 
 private:
     void Move() override { label.MoveTo({ pos.x, pos.y - label_offset }); }
